% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corphyloCpp.R
\name{boot_corphylo}
\alias{boot_corphylo}
\title{Internal function to do parametric bootstrapping from \code{corphylo} object}
\usage{
boot_corphylo(cp_obj, boot, boot_out = NULL, n_cores = 1)
}
\arguments{
\item{cp_obj}{\code{corphylo} object}

\item{boot}{Number of bootstrap replicates}

\item{boot_out}{Function to retrieve necessary info from corphylo object for each
bootstrap replicate. Defaults to \code{NULL}, which retrieves the correlation.}

\item{n_cores}{Number of cores to use. Defaults to 1.}
}
\value{
A vector/matrix with \code{B} items/rows of correlation estimates
}
\description{
Note: only works with \code{corphylo} with two input X variables.
}
\seealso{
\code{\link[ape]{corphylo}} \code{\link{corphylo_cpp}}
}
