\name{corphylo_cpp-package}
\alias{corphylo_cpp-package}
\alias{corphylo_cpp}
\docType{package}
\title{\packageTitle{corphylo_cpp}}
\description{\packageDescription{corphylo_cpp}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{corphylo_cpp}
  \packageIndices{corphylo_cpp}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{corphylo_cpp}
  
  Maintainer: \packageMaintainer{corphylo_cpp}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
