% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corphyloCpp.R
\name{corphylo_cpp}
\alias{corphylo_cpp}
\title{C++ version of ape::corphylo}
\usage{
corphylo_cpp(X, U = list(), SeM = NULL, phy = NULL, REML = TRUE,
  method = c("Nelder-Mead", "SANN"), constrain.d = FALSE, reltol = 10^-6,
  maxit.NM = 1000, maxit.SA = 1000, temp.SA = 1, tmax.SA = 1,
  verbose = FALSE)
}
\arguments{
\item{X}{a n x p matrix with p columns containing the values for the n taxa. Rows of X should have rownames matching the taxon names in phy.}

\item{U}{a list of p matrices corresponding to the p columns of X, with each matrix containing independent variables for the corresponding column of X. The rownames of each matrix within U must be the same as X, or alternatively, the order of values in rows must match those in X. If U is omitted, only the mean (aka intercept) for each column of X is estimated. If U[[i]] is NULL, only an intercept is estimated for X[, i]. If all values of U[[i]][j] are the same, this variable is automatically dropped from the analysis (i.e., there is no offset in the regression component of the model).}

\item{SeM}{a n x p matrix with p columns containing standard errors of the trait values in X. The rownames of SeM must be the same as X, or alternatively, the order of values in rows must match those in X. If SeM is omitted, the trait values are assumed to be known without error. If only some traits have mesurement errors, the remaining traits can be given zero-valued standard errors.}

\item{phy}{a phylo object giving the phylogenetic tree.  The rownames of phy must be the same as X, or alternatively, the order of values in rows must match those in X.}

\item{REML}{whether REML or ML is used for model fitting.}

\item{method}{in optim(), either Nelder-Mead simplex minimization or SANN (simulated annealing) minimization is used. If SANN is used, it is followed by Nelder-Mead minimization.}

\item{constrain.d}{if constrain.d is TRUE, the estimates of d are constrained to be between zero and 1. This can make estimation more stable and can be tried if convergence is problematic. This does not necessarily lead to loss of generality of the results, because before using corphylo, branch lengths of phy can be transformed so that the "starter" tree has strong phylogenetic signal.}

\item{reltol}{a control parameter dictating the relative tolerance for convergence in the optimization; see optim().}

\item{maxit.NM}{a control parameter dictating the maximum number of iterations in the optimization with Nelder-Mead minimization; see optim().}

\item{maxit.SA}{a control parameter dictating the maximum number of iterations in the optimization with SANN minimization; see optim().}

\item{temp.SA}{a control parameter dictating the starting temperature in the optimization with SANN minimization; see optim().}

\item{tmax.SA}{a control parameter dictating the number of function evaluations at each temperature in the optimization with SANN minimization; see optim().}

\item{verbose}{if TRUE, the model logLik and running estimates of the
    correlation coefficients and values of d are printed each iteration
    during optimization.}
}
\value{

}
\description{
For people in a hurry
}
\seealso{
\code{\link[ape]{corphylo}} \code{\link{boot_r}}
}
