% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corphyloCpp.R
\name{boot_r}
\alias{boot_r}
\title{Bootstrap Pearson r from \code{corphylo} object with two input X variables.}
\usage{
boot_r(cp_obj, B)
}
\arguments{
\item{cp_obj}{\code{corphylo} object}

\item{B}{Number of bootstrap replicates}
}
\value{
A vector of length \code{B} of correlation estimates
}
\description{
Bootstrap Pearson r from \code{corphylo} object with two input X variables.
}
\seealso{
\code{\link[ape]{corphylo}} \code{\link{corphylo_cpp}}
}
